dtc-live-tracker/fetcher/requirements.txt
requests
pymongo
protobuf
python-dotenv
